Fix for the music being queried twice, added MediaStoreUtil for media store related code. Added functions querySongsFromAlbum() and queryAvailableAlbums().
